Move file content management logic outside of providers under Chef::FileContentManagement for reusability in the future.
