+++ spec/unit/provider/cookbook_file_spec.rb
@@ -45,7 +45,7 @@ describe Chef::Provider::CookbookFile do
   end
 
   let(:content) do
-    content = mock('Chef::Provider::File::Content::CookbookFile')
+    content = mock('Chef::Provider::CookbookFile::Content')
   end
 
   it_behaves_like Chef::Provider::File
