+++ spec/unit/provider/file_spec.rb
@@ -29,7 +29,7 @@ describe Chef::Provider::File do
   end
 
   let(:content) do
-    content = mock('Chef::Provider::File::Content::File')
+    content = mock('Chef::Provider::File::Content')
   end
 
   let(:node) { double('Chef::Node') }
